+++ src/main.ts
@@ -1,9 +1,21 @@
 import * as vscode from 'vscode';
+/*
 import { TestHub, testExplorerExtensionId } from 'vscode-test-adapter-api';
 import { Log, TestAdapterRegistrar } from 'vscode-test-adapter-util';
 import { Z80UnitTestAdapter } from './adapter';
+*/
 
 export async function activate(context: vscode.ExtensionContext) {
+	const message: string = "Please de-install the 'Z80 Unit Tests' extension. It is no longer required. The functionality has been incorporated in DeZog 2.4. Please see https://github.com/maziac/DeZog/blob/main/documentation/Migration.md for more info.";
+	const msgDoc = "Open Migration.md";
+	vscode.window.showWarningMessage(message, msgDoc)
+		.then(result => {
+			vscode.env.openExternal(vscode.Uri.parse('https://github.com/maziac/DeZog/blob/main/documentation/Migration.md'));
+		});
+	// Also output to log
+	console.log(message);
+
+	/*
 	// Create a simple logger that can be configured with the configuration variables
 	// `z80UnitTestExplorer.logpanel` and `z80UnitTestExplorer.logfile`
 	let workspaceFolder;
@@ -27,4 +39,5 @@ export async function activate(context: vscode.ExtensionContext) {
 			log
 		));
 	}
+	*/
 }